Extract today's page and event calculation functions to
